+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Don't show "unrecognized parameter" warning for `inherit_mode` parameter to individual cop configurations. ([@maxh][])
 * [#6449](https://github.com/rubocop-hq/rubocop/pull/6449): Fix a false negative for `Layout/IndentationWidth` when setting `EnforcedStyle: rails` of `Layout/IndentationConsistency` and method definition indented to access modifier in a singleton class. ([@koic][])
 * [#6482](https://github.com/rubocop-hq/rubocop/issues/6482): Fix a false positive for `Lint/FormatParameterMismatch` when using (digit)$ flag. ([@koic][])
+* [#6250](https://github.com/rubocop-hq/rubocop/issues/6250): Skip Lint/UselessAccessModifier for `:new` method. ([@tejasbubane][])
 
 ## 0.60.0 (2018-10-26)